MNT: move tables and table views into widgets/tables
