+++ squirrel/widgets/page/pv_browser.py
@@ -9,11 +9,9 @@ from squirrel.model import PV
 from squirrel.permission_manager import PermissionManager
 from squirrel.utils import parse_csv_to_dict
 from squirrel.widgets.page.page import Page
-from squirrel.widgets.pv_browser_table import (PV_BROWSER_HEADER,
-                                               CSVTableModel,
-                                               PVBrowserFilterProxyModel,
-                                               PVBrowserTableModel)
-from squirrel.widgets.squirrel_table_view import SquirrelTableView
+from squirrel.widgets.tables import (PV_BROWSER_HEADER, CSVTableModel,
+                                     PVBrowserFilterProxyModel,
+                                     PVBrowserTableModel, SquirrelTableView)
 from squirrel.widgets.tag import TagDelegate, TagsWidget
 
 logger = logging.getLogger(__file__)
