+++ squirrel/widgets/page/snapshot_details.py
@@ -6,9 +6,8 @@ from qtpy import QtCore, QtWidgets
 from squirrel.client import Client
 from squirrel.model import Snapshot
 from squirrel.widgets.page.page import Page
-from squirrel.widgets.pv_table import PV_HEADER, PVTableModel
-from squirrel.widgets.snapshot_table import SnapshotTableModel
-from squirrel.widgets.squirrel_table_view import SquirrelTableView
+from squirrel.widgets.tables import (PV_HEADER, PVTableModel,
+                                     SnapshotTableModel, SquirrelTableView)
 
 
 class SnapshotDetailsPage(Page):
