+++ squirrel/widgets/tables/__init__.py
@@ -1,0 +1,7 @@
+from .pv_browser_table import (PV_BROWSER_HEADER, CSVTableModel,  # noqa
+                               PVBrowserFilterProxyModel, PVBrowserTableModel)
+from .pv_table import PV_HEADER, PVTableModel  # noqa
+from .snapshot_comparison_table import COMPARE_HEADER  # noqa
+from .snapshot_comparison_table import SnapshotComparisonTableModel  # noqa
+from .snapshot_table import SnapshotFilterModel, SnapshotTableModel  # noqa
+from .squirrel_table_view import SquirrelTableView  # noqa
