+++ squirrel/widgets/tables/snapshot_comparison_table.py
