+++ squirrel/widgets/tables/snapshot_table.py
