+++ squirrel/widgets/tables/squirrel_table_view.py
